Allow pixel coordinates as float

+++ opendm/gcp.py
@@ -39,7 +39,7 @@ class GCPFile:
             parts = entry.split()
             x, y, z, px, py, filename = parts[:6]
             extras = " ".join(parts[6:])
-            return GCPEntry(float(x), float(y), float(z), int(px), int(py), filename, extras)
+            return GCPEntry(float(x), float(y), float(z), float(px), float(py), filename, extras)
 
     def get_entry(self, n):
         if n < self.entries_count():
